Introduce a .topic.metadata.json file containing the activity information
Instead of hardocding the list of activities in JS we now have a "topic" which
has a name/description and a list of sub-activities. In our case the "library app"
is the topic.

This also moves the key used for routing into the activity metadata json file.

The paths in the topic file are relative to the location of the .json file itself.

See #22