Reload activities instead of hiding/showing them
Instead of loading them on demand and then hiding/showing them we add/remove
them from the DOM every time. The downside is that we lose state when switching between
activities, but we also don't have to update all loaded activities in the background even
if they are not used.

See #12