Dynamically create activity elements from the metadata files
Don't hardcode the element names in the templates and create everything based on the
information we get from the metadata files.

Because lit-element doesn't support creating tag names based on external data we have
to handle creating/removing elements and event bindings ourselves here.

We still hardcode the attributes we save/set which is the next thing we need to change.