Define a list of attributes and restore them between activities
We use a MutationObserver to observe attributes, which we then save in
session storage and set on the next activity.

The list of attributes is defined in the topic metadata and affects all acitvities
in that topic.

Because lit-element doesn't change the attributes when the associated properties changes
by default we have to set reflect=True for those properties to force attribute updates.