Work around CSS not being applied to slots in Safari
This resulted in the loading spinner being visible after the app was fully loaded.
Safari doesn't like class/style being set in the HTML or through CSS for slots and just ignores it.

This sets the display style in JS after the app is loaded and the slot is rendered and in the DOM.