Add a full page overlay for unsupported browsers
Add some feature detections for newer features we use and if they fail
show a full page overlay telling the user that the browser is too old or
unsupported.

Since we still send things to matomo also include a privacy notice and
an email link for the IT support in case the feature detection is broken.

Sadly we need to use eval() for feature detection so we have to loosen the
CSP rules.

See #56