Fix missing status translations in the order list
I remove the translations in fbb5e972f because i18next-scanner didn't find
the keys, but turns out the keys were generated at runtime, based on server
response data.

Replace this with a hardcoded list of keys for the known status values.
This makes i18next-scanner happy and also makes it easier to see what is translated
and what isn't.