Fix the sublibrary filtering for the book offer select
The format of the sublibrary IRI has changed and no longer includes the code,
so it was extracting an empty string. In addition to that the parameter
name in the API changed from "library" to "sublibrary" and now no longer
takes a code but the identifier for the library.

To fix this extract the ID from the IRI and pass as "sublibrary"