book-list: handle datePublished being freeform text
We need to move this from a ISO date string to a freeform text string
as the backend not always has data that can be converted losslessly.

In case it is an iso date, handle as before and show the year, but
in case it isn't show it as is.

For sorting we now need to handle random strings, so to keep things
working as before extract the earliest year-like number from the string.
This also has to handle uncertain year formats like "[20]14", see the
unit tests.